Force a specific weenie for new portal constructor

+++ Source/ACE/Command/Handlers/DebugCommands.cs
@@ -280,8 +280,7 @@ namespace ACE.Command.Handlers
             "Creates a portal in front of you.")]
         public static void CreatePortal(Session session, params string[] parameters)
         {
-            LandblockManager.AddObject(new Portal(ObjectType.Portal, new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.None), (ushort)PortalWeenie.HummingCrystal, session.Player.Location.InFrontOf(3.0f)));
-            // LandblockManager.AddObject(PortalObjectFactory.CreatePortal(1234, session.Player.Location.InFrontOf(3.0f), "Test Portal", PortalType.Purple));
+            LandblockManager.AddObject(new Portal(ObjectType.Portal, new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.None), PortalWeenie.HummingCrystal, session.Player.Location.InFrontOf(3.0f)));
         }
 
         /// <summary>
